Move composite tasks to separate module
